requirements.txt
flask==3.0.0
python-dotenv==1.0.0
slack-sdk==3.27.0

# Optional: for AI-powered action extraction (future)
# anthropic==0.18.0
